Move `decode_hex` to `utils.py`
