+++ etheno/utils.py
@@ -2,6 +2,13 @@ import socket
 from urllib.request import urlopen
 from urllib.error import HTTPError, URLError
 
+def decode_hex(data):
+    if data is None:
+        return None
+    if data[:2] == '0x':
+        data = data[2:]
+    return bytes.fromhex(data)
+
 def format_hex_address(addr):
     if isinstance(addr, int):
         addr = "%x" % addr
